# swagger.json
{
  "swagger": "2.0",
  "info": {
      "version": "1.0.0",
      "title": "My Project CRUD",
      "description": "My Project Application API",
      "license": {
          "name": "MIT",
          "url": "https://opensource.org/licenses/MIT"
      }
  },
  "host": "localhost:8008",
  "basePath": "/",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
  "/med/{id}": {
      "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of med",
            "type": "string"
          }
      ],
      "get": {
        "tags": ["Med by id"],
        "summary": "Get Med in system by id",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
  },
  "/med-delete/{id}": {
      "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of med to delete",
            "type": "string"
          }
      ],
      "delete": {
        "tags": ["Delete  med by id"],
        "summary": "Delete med in system by id",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
  },
  "/med-update/{id}": {
    "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "id of med",
          "type": "string"
        },
        {
          "name": "med",
          "in": "body",
          "description": "the med to create",
          "schema": {
            "type": "object",
            "properties": {
              "nume_medicament": {
                "type":"string"
              },
              "pret": {
                "type":"number"
              }
            }
          }
        }
    ],
      "patch": {
        "tags": ["Update med by id"],
        "summary": "Update med in system by id",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/med-create": {
    "parameters": [    
        {
          "name": "med",
          "in": "body",
          "description": "the med to create",
          "schema": {
            "type": "object",
            "properties": {
              "nume_medicament": {
                "type":"string"
              },
              "pret": {
                "type":"number"
              }
            }
          }
        }
    ],
    "put": {
      "tags": ["Create med"],
      "summary": "Create med in system",
      "responses": {
        "200": {
          "description": "OK"
        }
      }
    } 
  },
...
 }
}